requirements.txt
aniso8601==8.1.1
appnope==0.1.2
attrs==20.3.0
autobahn==21.1.1
Automat==20.2.0
backcall==0.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.4.3
dateparser==1.0.0
decorator==5.0.3
Flask==1.1.2
Flask-HTTPAuth==4.2.0
Flask-RESTful==0.3.8
flatten-json==0.1.13
google-api-core==1.26.0
google-auth==1.26.1
google-cloud-build==3.0.1
googleapis-common-protos==1.52.0
grpcio==1.35.0
gunicorn==20.0.4
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.12
jupyter-core==4.7.1
libcst==0.3.17
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.20.0
packaging==20.9
pandas==1.2.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
proto-plus==1.13.0
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-binance==0.7.10
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
pyzmq==22.0.3
regex==2020.11.13
requests==2.25.1
rsa==4.7
service-identity==18.1.0
six==1.15.0
tornado==6.1
traitlets==5.0.5
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
zope.interface==5.2.0
